feat: creation of firebase and simple layout on home page

+++ lib/core/models/wind_entity.dart
@@ -9,7 +9,7 @@ class WindEntity {
 
   factory WindEntity.fromJson(Map<String, dynamic> json) {
     return WindEntity(
-      speed: json['speed'],
+      speed: json['speed'].toDouble(),
       degress: json['deg'].toDouble(),
     );
   }
